+++ server/qradviser/models/response/search.py
@@ -1,5 +1,8 @@
+from typing import List, Union
 from uuid import UUID
 from pydantic import BaseModel
 class StopData(BaseModel):
     stop_id: UUID
     name: str
+    position: Union[List[float],None]
+    distance: float=None
